Create a dedicated test file to prepare for more unit tests
